Normed and moved functions
