+++ Makefile
@@ -21,6 +21,8 @@ CFILES		=	$(SRCDIR)/minirt.c \
 				$(SRCDIR)/scene/sphere.c \
 				$(SRCDIR)/scene/plane.c \
 				$(SRCDIR)/scene/cylinder.c \
+				$(SRCDIR)/movement/camera_movement.c \
+				$(SRCDIR)/movement/camera_rotation.c \
 				$(SRCDIR)/hierarchy/hierarchy.c \
 				$(SRCDIR)/hierarchy/hierarchy_utils.c \
 				$(SRCDIR)/hierarchy/hierarchy_sphere.c \
@@ -42,6 +44,7 @@ IFILES		=	$(INCLDIR)/minirt.h \
 				$(INCLDIR)/scene.h \
 				$(INCLDIR)/utils.h \
 				$(INCLDIR)/hierarchy.h \
+				$(INCLDIR)/movement.h \
 				$(INCLDIR)/vector.h
 
 LIBFT_DIR	=	./libft
