+++ include/movement.h
@@ -1,0 +1,27 @@
+#include <math.h>
+
+#define KEY_A 97
+#define KEY_D 100
+#define KEY_E 101
+#define KEY_Q 113
+#define KEY_R 114
+#define KEY_S 115
+#define KEY_W 119
+#define KEY_L 108
+#define KEY_J 106
+#define KEY_I 105
+#define KEY_K 107
+#define KEY_ESC 65307
+#define KEY_LEFT 65361
+#define KEY_UP 65362
+#define KEY_RIGHT 65363
+#define KEY_DOWN 65364
+#define MOVE_INTERVAL 0.35
+#define	ROTATE_ANGLE 0.03
+
+#define MAX_PITCH_RAD (M_PI / 2.0 - 0.01) // 89.4 max
+#define MIN_PITCH_RAD (-M_PI / 2.0 + 0.01)
+
+void	update_camera(t_core *core);
+void	rotate_camera_yaw(t_camera *cam, float angle);
+void	rotate_camera_pitch(t_camera *cam, float angle);
